+++ lib/src/controls/inputs/toggle_switch.dart
@@ -2,24 +2,23 @@ import 'package:fluent_ui/fluent_ui.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/rendering.dart';
 
-typedef ToggleSwitchThumbBuilder = Widget Function(
+typedef ToggleSwitchKnobBuilder = Widget Function(
   BuildContext context,
   Set<ButtonStates> states,
 );
 
-/// The toggle switch represents a physical switch that allows users to
-/// turn things on or off, like a light switch. Use toggle switch controls
-/// to present users with two mutually exclusive options (such as on/off),
-/// where choosing an option provides immediate results.
+/// The toggle switch represents a physical switch that allows users to turn
+/// things on or off, like a light switch. Use toggle switch controls to present
+/// users with two mutually exclusive options (such as on/off), where choosing
+/// an option provides immediate results.
 ///
-/// Use a toggle switch for binary operations that take effect right after
-/// the user flips the toggle switch
+/// Use a toggle switch for binary operations that take effect right after the
+/// user flips the toggle switch
 ///
 /// ![ToggleSwitch Preview](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/toggleswitches01.png)
 ///
-/// Think of the toggle switch as a physical power switch for a device: you
-/// flip it on or off when you want to enable or disable the action performed
-/// by the device.
+/// Think of the toggle switch as a physical power switch for a device: you flip
+/// it on or off when you want to enable or disable the action performed by the device.
 ///
 /// See also:
 ///
@@ -39,8 +38,8 @@ class ToggleSwitch extends StatefulWidget {
     this.content,
     this.leadingContent = false,
     this.semanticLabel,
-    this.thumb,
-    this.thumbBuilder,
+    this.knob,
+    this.knobBuilder,
     this.focusNode,
     this.autofocus = false,
   });
@@ -55,21 +54,22 @@ class ToggleSwitch extends StatefulWidget {
   /// If this callback is null, the switch is considered disabled.
   final ValueChanged<bool>? onChanged;
 
-  /// The thumb of the switch
+  /// The knob of the switch
   ///
-  /// [DefaultToggleSwitchThumb] is used by default
+  /// [DefaultToggleSwitchKnob] is used by default
   ///
   /// See also:
-  ///   * [thumbBuilder], which builds the thumb based on the current state
-  ///   * [DefaultToggleSwitchThumb], used when both [thumb] and [thumbBuilder] are null
-  final Widget? thumb;
+  ///
+  ///   * [knobBuilder], which builds the knob based on the current state
+  ///   * [DefaultToggleSwitchKnob], used when both [knob] and [knobBuilder] are null
+  final Widget? knob;
 
-  /// Build the thumb of the switch based on the current state
+  /// Build the knob of the switch based on the current state
   ///
   /// See also:
-  ///   * [thumb], a static thumb
-  ///   * [DefaultToggleSwitchThumb], used when both [thumb] and [thumbBuilder] are null
-  final ToggleSwitchThumbBuilder? thumbBuilder;
+  ///   * [knob], a static knob
+  ///   * [DefaultToggleSwitchKnob], used when both [knob] and [knobBuilder] are null
+  final ToggleSwitchKnobBuilder? knobBuilder;
 
   /// The style of the toggle switch
   final ToggleSwitchThemeData? style;
@@ -122,7 +122,9 @@ class _ToggleSwitchState extends State<ToggleSwitch> {
   Alignment? _alignment;
 
   void _handleAlignmentChanged(
-      Offset localPosition, double sliderGestureWidth) {
+    Offset localPosition,
+    double sliderGestureWidth,
+  ) {
     setState(() {
       _alignment = Alignment(
         (localPosition.dx / sliderGestureWidth).clamp(-1, 1),
@@ -142,30 +144,33 @@ class _ToggleSwitchState extends State<ToggleSwitch> {
       margin: style.margin,
       focusNode: widget.focusNode,
       onPressed: isDisabled ? null : () => widget.onChanged!(!widget.checked),
-      onHorizontalDragStart: (e) {
-        if (isDisabled) return;
-        _handleAlignmentChanged(e.localPosition, sliderGestureWidth);
-        setState(() => _dragging = true);
-      },
-      onHorizontalDragUpdate: (e) {
-        if (isDisabled) return;
-        _handleAlignmentChanged(e.localPosition, sliderGestureWidth);
-        if (!_dragging) setState(() => _dragging = true);
-      },
-      onHorizontalDragEnd: (e) {
-        if (isDisabled) return;
-        if (_alignment != null) {
-          if (_alignment!.x >= 0.5) {
-            widget.onChanged!(true);
-          } else {
-            widget.onChanged!(false);
-          }
-          setState(() {
-            _alignment = null;
-            _dragging = false;
-          });
-        }
-      },
+      onHorizontalDragStart: isDisabled
+          ? null
+          : (e) {
+              _handleAlignmentChanged(e.localPosition, sliderGestureWidth);
+              setState(() => _dragging = true);
+            },
+      onHorizontalDragUpdate: isDisabled
+          ? null
+          : (e) {
+              _handleAlignmentChanged(e.localPosition, sliderGestureWidth);
+              if (!_dragging) setState(() => _dragging = true);
+            },
+      onHorizontalDragEnd: isDisabled
+          ? null
+          : (e) {
+              if (_alignment != null) {
+                if (_alignment!.x >= 0.5) {
+                  widget.onChanged!(true);
+                } else {
+                  widget.onChanged!(false);
+                }
+                setState(() {
+                  _alignment = null;
+                  _dragging = false;
+                });
+              }
+            },
       builder: (context, states) {
         Widget child = AnimatedContainer(
           alignment: _alignment ??
@@ -180,28 +185,36 @@ class _ToggleSwitchState extends State<ToggleSwitch> {
           decoration: widget.checked
               ? style.checkedDecoration?.resolve(states)
               : style.uncheckedDecoration?.resolve(states),
-          child: widget.thumb ??
-              widget.thumbBuilder?.call(context, states) ??
-              DefaultToggleSwitchThumb(
+          child: widget.knob ??
+              widget.knobBuilder?.call(context, states) ??
+              DefaultToggleSwitchKnob(
                 checked: widget.checked,
                 style: style,
                 states: _dragging ? {ButtonStates.pressing} : states,
               ),
         );
         if (widget.content != null) {
-          child = Row(
-            mainAxisSize: MainAxisSize.min,
-            children: widget.leadingContent
-                ? [
-                    widget.content!,
-                    const SizedBox(width: 10.0),
-                    child,
-                  ]
-                : [
-                    child,
-                    const SizedBox(width: 10.0),
-                    widget.content!,
-                  ],
+          child = DefaultTextStyle.merge(
+            style: TextStyle(color: style.foregroundColor?.resolve(states)),
+            child: IconTheme.merge(
+              data: IconThemeData(
+                color: style.foregroundColor?.resolve(states),
+              ),
+              child: Row(
+                mainAxisSize: MainAxisSize.min,
+                children: widget.leadingContent
+                    ? [
+                        widget.content!,
+                        const SizedBox(width: 10.0),
+                        child,
+                      ]
+                    : [
+                        child,
+                        const SizedBox(width: 10.0),
+                        widget.content!,
+                      ],
+              ),
+            ),
           );
         }
         return Semantics(
@@ -216,8 +229,8 @@ class _ToggleSwitchState extends State<ToggleSwitch> {
   }
 }
 
-class DefaultToggleSwitchThumb extends StatelessWidget {
-  const DefaultToggleSwitchThumb({
+class DefaultToggleSwitchKnob extends StatelessWidget {
+  const DefaultToggleSwitchKnob({
     super.key,
     required this.checked,
     required this.style,
@@ -245,8 +258,8 @@ class DefaultToggleSwitchThumb extends StatelessWidget {
           (states.isHovering ? 2.0 : 0.0) +
           (states.isPressing ? 5.0 : 0),
       decoration: checked
-          ? style?.checkedThumbDecoration?.resolve(states)
-          : style?.uncheckedThumbDecoration?.resolve(states),
+          ? style?.checkedKnobDecoration?.resolve(states)
+          : style?.uncheckedKnobDecoration?.resolve(states),
     );
   }
 }
@@ -315,31 +328,48 @@ class ToggleSwitchTheme extends InheritedTheme {
 
 @immutable
 class ToggleSwitchThemeData with Diagnosticable {
-  final ButtonState<Decoration?>? checkedThumbDecoration;
-  final ButtonState<Decoration?>? uncheckedThumbDecoration;
+  /// The decoration of the knob when the switch is checked
+  final ButtonState<Decoration?>? checkedKnobDecoration;
 
+  /// The decoration of the knob when the switch is unchecked
+  final ButtonState<Decoration?>? uncheckedKnobDecoration;
+
+  /// The decoration of the switch when the it is checked
   final ButtonState<Decoration?>? checkedDecoration;
+
+  /// The decoration of the switch when the it is unchecked
   final ButtonState<Decoration?>? uncheckedDecoration;
 
+  /// The padding of the switch
   final EdgeInsetsGeometry? padding;
+
+  /// The margin of the switch
   final EdgeInsetsGeometry? margin;
 
+  /// The duration of the animation
   final Duration? animationDuration;
+
+  /// The curve of the animation
   final Curve? animationCurve;
 
+  /// The foreground color of the content of the switch
+  final ButtonState<Color?>? foregroundColor;
+
+  /// Creates a theme that can be used for [ToggleSwitchTheme]
   const ToggleSwitchThemeData({
     this.padding,
     this.margin,
     this.animationDuration,
     this.animationCurve,
-    this.checkedThumbDecoration,
-    this.uncheckedThumbDecoration,
+    this.checkedKnobDecoration,
+    this.uncheckedKnobDecoration,
     this.checkedDecoration,
     this.uncheckedDecoration,
+    this.foregroundColor,
   });
 
   factory ToggleSwitchThemeData.standard(FluentThemeData theme) {
-    final defaultThumbDecoration = BoxDecoration(
+    final defaultKnobDecoration = BoxDecoration(
       borderRadius: BorderRadius.circular(100),
     );
 
@@ -348,39 +378,46 @@ class ToggleSwitchThemeData with Diagnosticable {
     );
 
     return ToggleSwitchThemeData(
+      foregroundColor: ButtonState.resolveWith((states) {
+        return states.isDisabled ? theme.resources.textFillColorDisabled : null;
+      }),
       checkedDecoration: ButtonState.resolveWith((states) {
         return defaultDecoration.copyWith(
           color: ButtonThemeData.checkedInputColor(theme, states),
           border: Border.all(
-            color: Colors.transparent,
+            color: ButtonThemeData.checkedInputColor(theme, states),
           ),
         );
       }),
       uncheckedDecoration: ButtonState.resolveWith((states) {
         return defaultDecoration.copyWith(
-          color: states.isNone
-              ? theme.resources.subtleFillColorTransparent
-              : ButtonThemeData.uncheckedInputColor(theme, states),
+          color: ButtonState.forStates(
+            states,
+            disabled: theme.resources.controlAltFillColorDisabled,
+            pressed: theme.resources.controlAltFillColorQuarternary,
+            hovering: theme.resources.controlAltFillColorTertiary,
+            none: theme.resources.controlAltFillColorSecondary,
+          ),
           border: Border.all(
-            color: !states.isDisabled
-                ? theme.borderInputColor
-                : theme.brightness.isLight
-                    ? const Color.fromRGBO(0, 0, 0, 0.2169)
-                    : const Color.fromRGBO(255, 255, 255, 0.1581),
+            color: ButtonState.forStates(
+              states,
+              disabled: theme.resources.controlStrongFillColorDisabled,
+              none: theme.resources.controlStrongFillColorDefault,
+            ),
           ),
         );
       }),
       animationDuration: theme.fasterAnimationDuration,
-      animationCurve: theme.animationCurve,
-      checkedThumbDecoration: ButtonState.resolveWith((states) {
-        return defaultThumbDecoration.copyWith(
+      animationCurve: Curves.fastOutSlowIn,
+      checkedKnobDecoration: ButtonState.resolveWith((states) {
+        return defaultKnobDecoration.copyWith(
           color: states.isDisabled
               ? theme.resources.textOnAccentFillColorDisabled
               : theme.resources.textOnAccentFillColorPrimary,
         );
       }),
-      uncheckedThumbDecoration: ButtonState.resolveWith((states) {
-        return defaultThumbDecoration.copyWith(
+      uncheckedKnobDecoration: ButtonState.resolveWith((states) {
+        return defaultKnobDecoration.copyWith(
           color: states.isDisabled
               ? theme.resources.textFillColorDisabled
               : theme.resources.textFillColorSecondary,
@@ -400,14 +437,16 @@ class ToggleSwitchThemeData with Diagnosticable {
       animationCurve: t < 0.5 ? a?.animationCurve : b?.animationCurve,
       animationDuration: lerpDuration(a?.animationDuration ?? Duration.zero,
           b?.animationDuration ?? Duration.zero, t),
-      checkedThumbDecoration: ButtonState.lerp(a?.checkedThumbDecoration,
-          b?.checkedThumbDecoration, t, Decoration.lerp),
-      uncheckedThumbDecoration: ButtonState.lerp(a?.uncheckedThumbDecoration,
-          b?.uncheckedThumbDecoration, t, Decoration.lerp),
+      checkedKnobDecoration: ButtonState.lerp(a?.checkedKnobDecoration,
+          b?.checkedKnobDecoration, t, Decoration.lerp),
+      uncheckedKnobDecoration: ButtonState.lerp(a?.uncheckedKnobDecoration,
+          b?.uncheckedKnobDecoration, t, Decoration.lerp),
       checkedDecoration: ButtonState.lerp(
           a?.checkedDecoration, b?.checkedDecoration, t, Decoration.lerp),
       uncheckedDecoration: ButtonState.lerp(
           a?.uncheckedDecoration, b?.uncheckedDecoration, t, Decoration.lerp),
+      foregroundColor: ButtonState.lerp(
+          a?.foregroundColor, b?.foregroundColor, t, Color.lerp),
     );
   }
 
@@ -417,12 +456,13 @@ class ToggleSwitchThemeData with Diagnosticable {
       padding: style?.padding ?? padding,
       animationCurve: style?.animationCurve ?? animationCurve,
       animationDuration: style?.animationDuration ?? animationDuration,
-      checkedThumbDecoration:
-          style?.checkedThumbDecoration ?? checkedThumbDecoration,
-      uncheckedThumbDecoration:
-          style?.uncheckedThumbDecoration ?? uncheckedThumbDecoration,
+      checkedKnobDecoration:
+          style?.checkedKnobDecoration ?? checkedKnobDecoration,
+      uncheckedKnobDecoration:
+          style?.uncheckedKnobDecoration ?? uncheckedKnobDecoration,
       checkedDecoration: style?.checkedDecoration ?? checkedDecoration,
       uncheckedDecoration: style?.uncheckedDecoration ?? uncheckedDecoration,
+      foregroundColor: style?.foregroundColor ?? foregroundColor,
     );
   }
 
@@ -444,12 +484,13 @@ class ToggleSwitchThemeData with Diagnosticable {
         uncheckedDecoration,
       ))
       ..add(ObjectFlagProperty<ButtonState<Decoration?>?>.has(
-        'checkedThumbDecoration',
-        checkedThumbDecoration,
+        'checkedKnobDecoration',
+        checkedKnobDecoration,
       ))
       ..add(ObjectFlagProperty<ButtonState<Decoration?>?>.has(
-        'uncheckedThumbDecoration',
-        uncheckedThumbDecoration,
-      ));
+        'uncheckedKnobDecoration',
+        uncheckedKnobDecoration,
+      ))
+      ..add(DiagnosticsProperty('foregroundColor', foregroundColor));
   }
 }